backend/requirements.txt
celery==5.3.1
Django==3.2.20
djangorestframework==3.12.4
psycopg2-binary==2.9.7
kombu==5.3.1
redis==5.0.0
requests==2.31.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
